Warning
git-svn-id: https://svn.code.sf.net/p/bzflag/code/trunk/bzflag@8065 08b3d480-bf2c-0410-a26f-811ee3361c24

+++ src/3D/TextureManager.cxx
@@ -177,7 +177,7 @@ void TextureManager::setMaxFilter ( eTextureFilter filter )
   }
 
   // rebuild proc textures
-  for (int i = 0; i < countof(procLoader); i++) {
+  for (int i = 0; i < (int)countof(procLoader); i++) {
     procLoader[i].manager = this;
     procLoader[i].proc(procLoader[i]);
   }
